Improved Save/Load
The save and load carnival functions now use the NSSave and NSOpen panels provided by macOS, rather than import/export panels.

CarnivalSettings can now be accessed by a cog icon beside the navigation title as well as the context menu in the carnival list sidebar. This is the same CarnivalSettings used in the new carnival process. It also now provides the interface to set, and then change, the location of a carnival.

Data flow within the code has been changed in several scripts to keep things consistent and allow for better control and observation of the data. (Eg. Carnival Bindings)