Welcome screen
(Not yet implemented, but can be viewed)
The app now includes a welcome screen that will soon appear on app startup. This includes a title and version number, and points the user to three main options: New Carnival, Open Carnival, and View Documentation.

Creating or opening a carnival in the welcome screen will soon automatically open the main window if it is not already open.